requirements.txt
Pillow
reportlab
opencv-python
numpy
